add HomeController and Index action with view for admin dashboard.
add three DTOs including;
"AddUserRoleDTO" for assigning a role to an user,
"UserRolesDTO" for showing Roles of an user and
"UsersInRoleDTO"  for showing users of a role.

implementation of create, edit and index for "Roles" in RolesController.
add "UsersInRole" action for showing list of users of a role.

add "AddUserRole" action with view for assigning  a role to an user.
add "UserRoles" action and view to show a list of roles of an user.

modified layout od area admin and base layout.